Implement a `sync` API for working with global mutable state. (#107)
* Add IntelliJ workspace files to the .gitignore.

* Add a simple make_example script for Linux.

* Create a `sync` module with many GBA-specific sync utilties.

* Fix overflow error in debug mode in the hello_world crate.

* Fixes to DMA.

* Code cleanup for the sync module.

* Run rustfmt on new sync code.

* Fix up some names and documentation in the sync module.

* Add a few changes suggested by thomcc for the locks.

* Added needed compiler fences to `InitOnce::try_get`.
* Change the error in `RawMutex::raw_unlock` to better reflect the cause.

* Add a proper issue link to the __sync_synchronize hack.

* Disable interrupts during `InitOnce::try_get`.

* Fix some bad wording in the comments for `InitOnce::try_get`

* Use the new target in `cfg` checks to see if we're on GBA.

* Change registers used for transfer_align4_arm for the different target.

* Cleanup on sync_api changes for the target change.

+++ examples/hello_world.rs
@@ -52,8 +52,8 @@ fn main(_argc: isize, _argv: *const *const u8) -> isize {
     let this_frame_keys = read_key_input();
 
     // adjust game state and wait for vblank
-    px = px.wrapping_add(2 * this_frame_keys.x_tribool() as usize);
-    py = py.wrapping_add(2 * this_frame_keys.y_tribool() as usize);
+    px = px.wrapping_add((2 * this_frame_keys.x_tribool() as i32) as usize);
+    py = py.wrapping_add((2 * this_frame_keys.y_tribool() as i32) as usize);
     if this_frame_keys.l() {
       color = Color(color.0.rotate_left(5));
     }
